Handle loading in a remote Markdown file
This is just a very rough cut; lots of extra stuff needs to happen here.
This also exposes a couple of things that need to be worked out. For
example: the MarkdownViewer only does navigation tracking (history stack
thing) iff you're giving it a path on the local filesystem. When pulling
from the web I need to just view the text.

The obvious thing to do here is to store in a temp file and then navigate to
that. Only... temp file; the clue is in the name! So ideally, at some point,
I need to wrap up the navigation class in something that can handle there
being URLs in the stack and do the right thing.