Move the help text into its own file
It's growing a bit, so this makes sense. There is, of course, an argument to
be made for having it as an external markdown document; but I don't really
want to have things that the application needs be in external "data" files.
Here I have one eye on possible options for deploying in some sort of
bundled/executable way and such files can be problematic.