Try and handle permission errors when filtering
This is more for Windows than anything else, and I'm not sure if this is the
right place to do it yet. This is an experimental commit so that I can then
pull it in over on Windows and test it (I really can't handle trying to
develop in Windows).